Add files via upload
TECH-KRISHI 

Agriculture is the backbone of our Indian economy, which plays an important role in increasing our country's economic development. The main focus of this Web application is on the agricultural sector because the farmers have the highest contribution to our country's GDP.
                This application provides the information to the farmer in real time and introduce farmers to the most to suitable crops to be sown along with fertilizer recommendation and plant disease detection, thereby helping farmers to achieve greater productivity with reduced loss.
 Renting Tools :At present, farmers need to travel to a place to borrow all the essential needs, which is a tiresome and not a cost effective work. So a smart digital farming is listed as the highest ranking technology opportunity in the latest Global Opportunity report in terms of its expected positive impact on society
Climate: Small changes in temperature and rainfall have significant effects on the quality of crops. This application helps in detecting suitable weather for particular crop.
Expert Assistance : This includes contacts and feedbacks